Update code to make use of the cleanup that has been done on search tasks.
Earlier in the 5.x series Matt Wire did a big cleanup on search tasks & we can use this now. I'm not sure exactly which release
but 5.6 is the current one so just picking that seems fine. Since this was not previously on the extensions directly I think
the current 'user' is PTP & I don't expect Jamie to have any difficulty navigating not deploying the version
with the cleaned up code on old Civi releases (I think 5.3 actually does have the fixes in it which is possibly what you
are on