Run civix.upgrade
This removes the code now-replaced by mixins. Rather than commit the mixin
folder I incremented the version, leaving the older version still in play
for older releases. The mixin folder is required for
 releases <=v5.44 - which is older than the various security updates
so reducing code for the future feels more useful than
preserving one release for before & after that

https://civicrm.org/blog/totten/civix-v2205-how-remove-million-lines-extra-code